# openapi_spec.yaml
openapi: 3.0.3
info:
  title: CliniSync / ShebaSync API Contract
  description: Official locked REST API specification for Frontend (Next.js & Flutter)
    and Backend (FastAPI).
  version: 1.0.0
servers:
- url: https://api.clinisync.com/api/v1
  description: Production Server
- url: http://localhost:8000/api/v1
  description: Local Development Server
paths:
  /auth/register:
    post:
      summary: Register patient account
      tags:
      - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ErrorResponse'
  /auth/login:
    post:
      summary: Authenticate user
      tags:
      - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/ErrorResponse'
  /auth/refresh:
    post:
      summary: Renew access token
      tags:
      - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Token renewed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  tokenType:
                    type: string
                    example: Bearer
                  expiresIn:
                    type: integer
                    example: 900
  /users/me:
    get:
      summary: Get own profile
      tags:
      - Users
      security:
      - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      summary: Update own profile
      tags:
      - Users
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                timezone:
                  type: string
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /doctors:
    get:
      summary: List doctors
      tags:
      - Doctors
      parameters:
      - name: specialization
        in: query
        schema:
          type: string
      - name: search
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Doctor list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Doctor'
    post:
      summary: Onboard doctor (Admin only)
      tags:
      - Doctors
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDoctorRequest'
      responses:
        '201':
          description: Doctor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
...
  /appointments/doctor/{doctorId}:
    get:
      summary: Doctor daily queue
      tags:
      - Appointments
      security:
      - BearerAuth: []
      parameters:
      - name: doctorId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: date
        in: query
        required: true
        schema:
          type: string
          format: date
      responses:
        '200':
          description: Daily queue list
...
  /devices/fcm-token:
    post:
      summary: Register or update FCM push device token
      tags:
      - Notifications
      security:
      - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - fcmToken
              - deviceType
              properties:
                fcmToken:
                  type: string
                deviceType:
                  type: string
                  enum:
                  - android
                  - ios
                  - web
                deviceModel:
                  type: string
      responses:
        '200':
          description: Token registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  deviceId:
                    type: string
                    format: uuid
                  fcmToken:
                    type: string
                  deviceType:
                    type: string
                  isActive:
                    type: boolean
                  updatedAt:
                    type: string
                    format: date-time
  /notifications:
    get:
      summary: Get paginated in-app notifications
      tags:
      - Notifications
      security:
      - BearerAuth: []
      parameters:
      - name: page
        in: query
        schema:
          type: integer
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          default: 20
      - name: unreadOnly
        in: query
        schema:
          type: boolean
          default: false
      responses:
        '200':
          description: Notification feed
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NotificationItem'
                  unreadCount:
                    type: integer
                  total:
                    type: integer
                  page:
                    type: integer
                  pageSize:
                    type: integer
  /notifications/{id}/read:
    patch:
      summary: Mark individual notification as read
      tags:
      - Notifications
      security:
      - BearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Notification marked read
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  isRead:
                    type: boolean
                  readAt:
                    type: string
                    format: date-time
  /notifications/mark-all-read:
    post:
      summary: Mark all notifications as read
      tags:
      - Notifications
      security:
      - BearerAuth: []
      responses:
        '200':
          description: All notifications marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  markedCount:
                    type: integer
  /auth/forgot-password:
    post:
      summary: Request password reset link
      tags:
      - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Reset token sent if email exists
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/reset-password:
    post:
      summary: Reset password using signed token
      tags:
      - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - token
              - newPassword
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    RegisterRequest:
      type: object
      required:
      - name
      - email
      - password
      properties:
        name:
          type: string
          example: Rahim Ahmed
        email:
          type: string
          format: email
          example: rahim@example.com
        password:
          type: string
          example: SecurePass123!
        timezone:
          type: string
          example: Asia/Dhaka
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
          - patient
          - doctor
          - admin
        timezone:
          type: string
        createdAt:
          type: string
          format: date-time
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        tokens:
          type: object
          properties:
            accessToken:
              type: string
            refreshToken:
              type: string
            tokenType:
              type: string
              example: Bearer
            expiresIn:
              type: integer
              example: 900
    Doctor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
          example: Prof. Dr. Karim Chowdhury
        specialization:
          type: string
          example: Cardiology
        degrees:
          type: string
          example: MBBS, FCPS (Cardiology), MD, FACC
        bmdcNumber:
          type: string
          example: 'BMDC Reg # A-28491'
        designation:
          type: string
          example: Senior Consultant & Professor
        facilityName:
          type: string
          example: Popular Diagnostic Centre, Dhanmondi
        chamberRoom:
          type: string
          example: 'Chamber #402, Level 4'
        profilePhotoUrl:
          type: string
          format: uri
        bio:
          type: string
        consultationFee:
          type: number
          example: 1200.0
        followupFee:
          type: number
          example: 800.0
        experienceYears:
          type: integer
          example: 18
        rating:
          type: number
          example: 4.9
    CreateDoctorRequest:
      type: object
      required:
      - name
      - email
      - password
      - specialization
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        specialization:
          type: string
        bio:
          type: string
        timezone:
          type: string
...
    ErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - status
          properties:
            code:
              type: string
              example: SLOT_CONFLICT
            message:
              type: string
              example: This slot was just booked by another patient.
            status:
              type: integer
              example: 409
            details:
              type: object
    NotificationItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        notificationType:
          type: string
          enum:
          - booking_confirmed
          - queue_update
          - doctor_break
          - reminder_24h
          - reminder_1h
          - cancelled
          - system
        appointmentId:
          type: string
          format: uuid
        isRead:
          type: boolean
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time